Refactor: 컴포넌트 분리 - Main컴포넌트에서 MainRightHeader, Recommend, Story 컴포넌트 분리 Modify: 불필요한 주석 삭제
