+++ src/pages/jinwookim/Main/Main.scss
@@ -1,8 +1,8 @@
+/* stylelint-disable string-quotes */
 @import "../../../styles/common";
 @import "../../jinwookim/Main/Feeds";
 
 .main-instagram-text {
-  /* stylelint-disable-next-line string-quotes */
   font-family: "Lobster", sans-serif;
   font-size: 20px;
 }
