+++ src/pages/jinwookim/Main/MainRightHeader.js
@@ -1,0 +1,14 @@
+import React from "react";
+const MainRightHeader = () => {
+  return (
+    <div className="main-right-header">
+      <img
+        className="main-right-icon"
+        src="https://images.cobot.me/uploads/space/logo/space-wecode/pdf_ab211233b6576dbb0f8b5826447eeac61e2a833a99ac5d788fbc1a174c3c6ce5.png"
+        alt="메인 오른쪽 아이콘"
+      />
+      <span style={{ fontWeight: "bold" }}>wecode_bootcamp</span>
+    </div>
+  );
+};
+export default MainRightHeader;
